requirements.txt
sentence-transformers
faiss-cpu #faiss-gpu
requests